+++ src/mmwave/model/mmwave-channel-raytracing.h
@@ -102,8 +102,8 @@ private:
                                                    Ptr<const MobilityModel> a,
                                                    Ptr<const MobilityModel> b) const;
 
-  complex2DVector_t GenSpatialMatrix (uint64_t traceIndex, uint8_t* antennaNum, bool bs) const;
-  complexVector_t GenSinglePath (double hAngle, double vAngle, uint8_t* antennaNum) const;
+  complex2DVector_t GenSpatialMatrix (uint64_t traceIndex, uint16_t* antennaNum, bool bs) const;
+  complexVector_t GenSinglePath (double hAngle, double vAngle, uint16_t* antennaNum) const;
   complexVector_t CalcBeamformingVector (complex2DVector_t SpatialMatrix, doubleVector_t powerFraction) const;
   Ptr<SpectrumValue> GetChannelGain (Ptr<const SpectrumValue> txPsd, Ptr<mmWaveBeamFormingTraces> bfParams, double speed) const;
   double GetSystemBandwidth () const;
